+++ packages/cosmwasm/src/signingcosmwasmclient.spec.ts
@@ -1,13 +1,28 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { Sha256 } from "@cosmjs/crypto";
 import { toHex } from "@cosmjs/encoding";
-import { AuthExtension, coin, coins, LcdClient, Secp256k1Wallet, setupAuthExtension } from "@cosmjs/sdk38";
+import {
+  AuthExtension,
+  coin,
+  coins,
+  LcdClient,
+  MsgDelegate,
+  Secp256k1Wallet,
+  setupAuthExtension,
+} from "@cosmjs/sdk38";
 import { assert } from "@cosmjs/utils";
 
 import { isPostTxFailure, PrivateCosmWasmClient } from "./cosmwasmclient";
 import { setupWasmExtension, WasmExtension } from "./lcdapi/wasm";
 import { SigningCosmWasmClient, UploadMeta } from "./signingcosmwasmclient";
-import { alice, getHackatom, makeRandomAddress, pendingWithoutWasmd, unused } from "./testutils.spec";
+import {
+  alice,
+  getHackatom,
+  makeRandomAddress,
+  pendingWithoutWasmd,
+  unused,
+  validatorAddress,
+} from "./testutils.spec";
 
 const httpUrl = "http://localhost:1317";
 
@@ -325,6 +340,29 @@ describe("SigningCosmWasmClient", () => {
       const after = await client.getAccount(beneficiaryAddress);
       assert(after);
       expect(after.balance).toEqual(transferAmount);
+    });
+  });
+
+  describe("signAndPost", () => {
+    it("works", async () => {
+      pendingWithoutWasmd();
+      const wallet = await Secp256k1Wallet.fromMnemonic(alice.mnemonic);
+      const client = new SigningCosmWasmClient(httpUrl, alice.address0, wallet);
+
+      const msg: MsgDelegate = {
+        type: "cosmos-sdk/MsgDelegate",
+        value: {
+          delegator_address: alice.address0,
+          validator_address: validatorAddress,
+          amount: coin(1234, "ustake"),
+        },
+      };
+      const fee = {
+        amount: coins(2000, "ucosm"),
+        gas: "120000", // 120k
+      };
+      const result = await client.signAndPost([msg], fee, "Use your power wisely");
+      assert(!isPostTxFailure(result));
     });
   });
 });